done fixes requested in git
